move functions into a separate file, improve comments
