+++ pkg/collections/collection.go
@@ -3,23 +3,23 @@
 // license that can be found in the LICENSE file.
 
 // Package collections provides an implementation of a generic Collection.
-// Collection wraps an underlying Go slice and provides convenient methods.
+// A collection wraps an underlying Go slice and provides convenience methods.
 // The design of this package is inspired by Scala's collections library.
-// This file contains the constructor, and basic introspection methods.
 package collections
 
 import (
 	"fmt"
 	"slices"
+
+	"github.com/charbz/gophers/pkg/utils"
 )
 
-// Collection represents a generic collection of elements of type T.
 type Collection[T any] struct {
 	elements []T
 }
 
 // NewCollection is a constructor for collections it takes
-// a variadic argument of slices of type T as input and converts
+// a variadic argument of slices of any type as input and converts
 // them to a single flat collection.
 func NewCollection[T any](s ...[]T) Collection[T] {
 	if len(s) == 0 {
@@ -28,27 +28,26 @@ func NewCollection[T any](s ...[]T) Collection[T] {
 	return Collection[T]{elements: slices.Concat(s...)}
 }
 
+// String implements the Stringer interface so fmt
+// can print the underlying slice
 func (c *Collection[T]) String() string {
 	return fmt.Sprintf("%v", c.elements)
 }
 
 // zero returns the zero value of the underlying slice
-// this is only used internally when an operation must
-// return an empty collection.
+// used internally when an operation must return an empty collection.
 func (c *Collection[T]) zero() *Collection[T] {
 	var z Collection[T]
 	return &z
 }
 
-// zeroT returns the zero value of
-// the underlying slice's type.
+// zeroT returns the zero value of the underlying slice's type.
 func (c *Collection[T]) zeroT() T {
 	var z T
 	return z
 }
 
-// At returns the element at the specified
-// index in the Collection.
+// At returns the element at the specified index.
 func (c *Collection[T]) At(index int) T {
 	return c.elements[index]
 }
@@ -58,14 +57,13 @@ func (c *Collection[T]) ToSlice() []T {
 	return c.elements
 }
 
-// IsEmpty returns true if the Collection contains
-// 0 elements.
+// IsEmpty returns true if the Collection contains 0 elements.
 func (c *Collection[T]) IsEmpty() bool {
 	return len(c.elements) == 0
 }
 
-// NonEmpty returns true if the Collection contains
-// at least 1 element.
+// NonEmpty returns true if the Collection contains at
+// least 1 element.
 func (c *Collection[T]) NonEmpty() bool {
 	return len(c.elements) > 0
 }
@@ -75,9 +73,8 @@ func (c *Collection[T]) Length() int {
 	return len(c.elements)
 }
 
-// Head returns the first element in the Collection
-// and nil error if non-empty, or zero value and
-// error otherwise.
+// Head returns the first element in the Collection and nil
+// error if non-empty, otherwise it returns zero value and error.
 func (c *Collection[T]) Head() (T, error) {
 	if c.IsEmpty() {
 		return c.zeroT(), EmptyCollectionError
@@ -86,8 +83,7 @@ func (c *Collection[T]) Head() (T, error) {
 }
 
 // Last returns the final element in the Collection
-// and nil error if non-empty, or zero value and
-// error otherwise.
+// and nil error if non-empty, or zero value and error otherwise.
 func (c *Collection[T]) Last() (T, error) {
 	if c.IsEmpty() {
 		return c.zeroT(), EmptyCollectionError
@@ -95,8 +91,7 @@ func (c *Collection[T]) Last() (T, error) {
 	return c.elements[len(c.elements)-1], nil
 }
 
-// Tail returns a slice containing all elements except the first.
-// If the Collection is empty, returns an empty slice.
+// Tail returns a collection containing all elements except the first.
 func (c *Collection[T]) Tail() *Collection[T] {
 	if c.IsEmpty() {
 		return c
@@ -106,8 +101,7 @@ func (c *Collection[T]) Tail() *Collection[T] {
 	}
 }
 
-// Init returns a slice containing all elements except the last.
-// If the Collection is empty, returns an empty slice.
+// Init returns a collection containing all elements except the last.
 func (c *Collection[T]) Init() *Collection[T] {
 	if c.IsEmpty() {
 		return c
@@ -117,54 +111,81 @@ func (c *Collection[T]) Init() *Collection[T] {
 	}
 }
 
-// Take returns a slice containing the first q elements.
-// If q is negative or zero, returns an empty slice.
-// If q exceeds the Collection length, returns all elements.
-func (c *Collection[T]) Take(q int) *Collection[T] {
-	if q <= 0 {
+// Take returns a collection containing the first n elements.
+func (c *Collection[T]) Take(n int) *Collection[T] {
+	if n <= 0 {
 		return c.zero()
 	}
 	return &Collection[T]{
-		c.elements[0:min(q, c.Length())],
+		c.elements[0:min(n, c.Length())],
 	}
 }
 
-// TakeRight returns a slice containing the last q elements.
-// If q is negative or zero, returns an empty slice.
-// If q exceeds the Collection length, returns all elements.
-func (c *Collection[T]) TakeRight(q int) *Collection[T] {
-	if q <= 0 {
+// TakeRight returns a collection containing the last n elements.
+func (c *Collection[T]) TakeRight(n int) *Collection[T] {
+	if n <= 0 {
 		return c.zero()
 	}
 	return &Collection[T]{
-		c.elements[max(c.Length()-q, 0):],
+		c.elements[max(c.Length()-n, 0):],
 	}
 }
 
-// Drop returns a slice with the first q elements removed.
-// If q is negative or zero, returns all elements.
-// If q exceeds the Collection length, returns an empty slice.
-func (c *Collection[T]) Drop(q int) *Collection[T] {
-	if q <= 0 {
+// Drop returns a collection with the first n elements removed.
+func (c *Collection[T]) Drop(n int) *Collection[T] {
+	if n <= 0 {
 		return c
-	} else if q >= c.Length() {
+	} else if n >= c.Length() {
 		return c.zero()
 	}
 	return &Collection[T]{
-		c.elements[q:],
+		c.elements[n:],
 	}
 }
 
-// DropRight returns a slice with the last q elements removed.
-// If q is negative or zero, returns all elements.
-// If q exceeds the Collection length, returns an empty slice.
-func (c *Collection[T]) DropRight(q int) *Collection[T] {
-	if q <= 0 {
+// DropRight returns a collection with the last n elements removed.
+func (c *Collection[T]) DropRight(n int) *Collection[T] {
+	if n <= 0 {
 		return c
-	} else if q >= c.Length() {
+	} else if n >= c.Length() {
 		return c.zero()
 	}
 	return &Collection[T]{
-		c.elements[0 : c.Length()-q],
+		c.elements[0 : c.Length()-n],
 	}
+}
+
+// Filter takes a filtering function as input and returns
+// the resulting collection after applying the filter to each element.
+func (c *Collection[T]) Filter(f func(T) bool) *Collection[T] {
+	return &Collection[T]{
+		utils.Filter(c.elements, f),
+	}
+}
+
+// FilterNot takes a filtering function as input and returns
+// the collection containing all elements that do not satisfy
+// the filtering function provided.
+func (c *Collection[T]) FilterNot(f func(T) bool) *Collection[T] {
+	return &Collection[T]{
+		utils.FilterNot(c.elements, f),
+	}
+}
+
+// Partition takes a partitioning function as input and returns
+// two collections, the first one containing the elements that
+// match the partitioning condition, and the other one contains
+// the rest of the elements.
+func (c *Collection[T]) Partition(f func(T) bool) (*Collection[T], *Collection[T]) {
+	left, right := utils.Partition(c.elements, f)
+	return &Collection[T]{left}, &Collection[T]{right}
+}
+
+// ForEach takes a function as input and applies the
+// function to each element in the collection.
+func (c *Collection[T]) ForEach(f func(T)) *Collection[T] {
+	for v := range c.Values() {
+		f(v)
+	}
+	return c
 }
